Use scope_key in summary API operation
The user can define any label of their metrics as the scope key,
the scope_key parameter exists. However, it is not taken in account for
the summary operation, where project_id is hardcoded instead of using
the appropiate scope_key. This fixes that behaviour and extends the
scope_key parameter usage to the summary operation

Change-Id: Iaa092d33f6cbd02b415116cdd4337ef9299c429e
Signed-off-by: Juan Larriba <jlarriba@redhat.com>